Реализовано добавление в базу данных после загрузки, реализованы мапперу, добавлены дагер иньекции

+++ app/src/main/java/ru/guzeyst/a65appstest/domain/model/Specialty.kt
@@ -2,5 +2,5 @@ package ru.guzeyst.a65appstest.domain.model
 
 data class Specialty(
     val name: String,
-    val specialty_id: Int
+    val specialty_id: Long
 )
